requirements.txt
# CLIP Embeddings API - Requirements
# Python 3.8+ required

# Core Flask framework
flask==2.3.3
flask-cors==4.0.0

# PyTorch for deep learning
torch==2.0.1
torchvision==0.15.2

# CLIP model dependencies
git+https://github.com/openai/CLIP.git
ftfy==6.1.1
regex==2023.8.8
tqdm==4.66.1

# Image processing
pillow==10.0.0

# Numerical computing
numpy==1.24.3

# HTTP requests
requests==2.31.0

# Vector database client
qdrant-client==1.4.0

# Additional utilities
python-dateutil==2.8.2

# For loading environment variables from a .env file
python-dotenv